many: move TestInterface to ifacetest package
