+++ interfaces/mount/backend_test.go
@@ -33,6 +33,7 @@ import (
 	"github.com/snapcore/snapd/dirs"
 	"github.com/snapcore/snapd/interfaces"
 	"github.com/snapcore/snapd/interfaces/backendtest"
+	"github.com/snapcore/snapd/interfaces/ifacetest"
 	"github.com/snapcore/snapd/interfaces/mount"
 	"github.com/snapcore/snapd/osutil"
 )
@@ -44,7 +45,7 @@ func Test(t *testing.T) {
 type backendSuite struct {
 	backendtest.BackendSuite
 
-	iface2 *interfaces.TestInterface
+	iface2 *ifacetest.TestInterface
 }
 
 var _ = Suite(&backendSuite{})
@@ -57,7 +58,7 @@ func (s *backendSuite) SetUpTest(c *C) {
 	c.Assert(err, IsNil)
 
 	// add second iface so that we actually test combining snippets
-	s.iface2 = &interfaces.TestInterface{InterfaceName: "iface2"}
+	s.iface2 = &ifacetest.TestInterface{InterfaceName: "iface2"}
 	err = s.Repo.AddInterface(s.iface2)
 	c.Assert(err, IsNil)
 }
